app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "oxapy",
    "python-dotenv",
    "sqlalchemy-libsql",
]
